refactor: 패키지 구조 MVC 형태로 수정
