+++ src/main/java/lotto/LottoMain.java
@@ -1,0 +1,9 @@
+package lotto;
+
+import lotto.controller.LottoController;
+
+public class LottoMain {
+    public static void main(String[] args) {
+        LottoController.startLotto();
+    }
+}
