+++ src/main/java/lotto/controller/LottoController.java
@@ -1,0 +1,7 @@
+package lotto.controller;
+
+public class LottoController {
+    public static void startLotto() {
+
+    }
+}
