+++ src/main/java/lotto/strategy/PickNumberStrategy.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.strategy;
 
 import java.util.List;
+import lotto.domain.LottoNumber;
 
 public interface PickNumberStrategy {
     List<LottoNumber> pickLottoNumbers();
